Move data into fixtures.
