+++ test/units/module_utils/acme/fixtures/csr_2.txt
@@ -1,0 +1,78 @@
+Certificate Request:
+    Data:
+        Version: 1 (0x0)
+        Subject: 
+        Subject Public Key Info:
+            Public Key Algorithm: rsaEncryption
+                RSA Public-Key: (4096 bit)
+                Modulus:
+                    00:db:f5:57:b8:03:b2:1e:fa:3b:ff:dd:f7:07:25:
+                    05:c5:ba:90:da:56:79:1e:9e:02:08:33:4e:1c:1b:
+                    31:98:dc:3b:b1:85:1d:0b:01:64:26:46:19:57:72:
+                    4b:7e:ee:85:e8:dc:46:b2:01:76:d0:c3:59:78:68:
+                    87:70:f3:f8:64:31:0f:ee:86:bb:00:04:17:22:f5:
+                    17:73:1f:4f:4b:f5:1c:7a:0e:f1:f1:3b:e0:6e:66:
+                    0f:af:99:03:61:d1:0d:e5:3d:8d:ff:01:3f:25:56:
+                    1c:1b:4d:6c:29:89:19:29:52:34:e1:5e:8d:60:a6:
+                    c2:fb:0b:01:14:31:a8:50:6e:4f:24:63:82:8b:38:
+                    06:42:f2:58:ca:99:46:e9:2b:cc:ed:4f:cf:b2:01:
+                    86:e1:6b:07:ad:3c:b7:e0:06:e2:f1:a7:3d:76:4d:
+                    33:bc:5e:7e:1c:ce:f1:b3:ce:89:ea:53:d1:84:b6:
+                    11:13:e4:16:63:b2:f7:b4:a7:c1:31:9d:2b:29:30:
+                    65:83:3d:a0:ab:a4:62:11:38:67:ec:8e:69:31:69:
+                    24:c0:4d:5f:00:c2:23:94:cc:9e:1b:5c:e9:e2:6b:
+                    7f:9b:0e:14:7f:76:9f:61:3d:f9:85:8c:a1:a8:54:
+                    3a:24:f0:1b:a5:d5:e4:77:d2:a7:24:a9:c0:e4:54:
+                    69:32:13:41:1b:0f:00:cf:9b:98:af:5f:d6:30:7a:
+                    72:d7:c5:57:0a:5a:f2:e5:bf:87:ca:52:81:b1:e0:
+                    66:f9:1e:0f:ab:3a:ac:52:e1:f5:34:36:e2:dd:2f:
+                    ad:19:a1:9d:8e:ff:00:c1:60:6e:02:68:05:30:6a:
+                    d2:5d:25:11:74:b8:14:21:8a:0c:6a:90:3a:a4:b0:
+                    b3:e7:87:47:fa:47:03:8b:98:bf:c7:78:f7:c4:90:
+                    97:d7:d8:b1:3f:dc:ed:48:f7:fc:99:cc:49:15:4d:
+                    f7:8d:b2:41:41:42:cf:02:df:d3:f5:b3:58:28:b6:
+                    1d:3a:8f:33:c0:8f:87:9a:f2:88:9d:f4:18:03:76:
+                    79:bd:36:eb:f9:e8:c9:33:97:9b:8e:fd:4e:0f:68:
+                    f3:a4:f0:f6:a7:33:13:48:f4:09:d7:11:9e:be:89:
+                    df:7e:d1:18:74:20:e7:c3:d9:ee:3f:a1:9d:49:8a:
+                    c2:0c:9c:97:2b:e3:13:5d:19:14:18:13:a0:c1:92:
+                    9f:36:29:e5:47:91:70:e1:59:dd:b9:ca:e8:b3:ac:
+                    18:2e:f9:98:8b:f3:d8:e4:20:49:07:8e:e5:54:00:
+                    48:49:ce:f6:e8:72:80:dc:a8:f7:5b:0a:e0:1b:55:
+                    ae:89:68:8a:df:19:09:af:36:2e:5b:07:48:28:ff:
+                    98:cc:cf
+                Exponent: 65537 (0x10001)
+        Attributes:
+        Requested Extensions:
+            X509v3 Subject Alternative Name: 
+                DNS:ansible.com, IP Address:127.0.0.1, IP Address:0:0:0:0:0:0:0:1, IP Address:2001:D88:AC10:FE01:0:0:0:0, IP Address:2001:1234:5678:ABCD:9876:5432:10FE:DCBA
+    Signature Algorithm: sha256WithRSAEncryption
+         45:46:e0:0d:cd:54:5c:cd:ef:a2:ba:bd:18:7a:f8:ed:60:85:
+         c3:6b:4d:c6:a1:65:a4:82:72:ee:13:cb:bc:7a:94:cc:db:6e:
+         3c:76:e1:b7:4e:6b:d5:bf:15:8f:37:88:85:ae:f1:5c:64:8b:
+         cd:b1:eb:c1:a3:e6:7b:91:72:76:e2:6d:d8:95:db:57:bf:07:
+         c0:62:60:99:d3:ad:9a:a5:2a:21:cb:38:28:dd:0d:0a:7c:37:
+         a1:c0:b7:28:24:f7:b9:6e:1f:a3:6d:b8:09:54:61:af:26:e8:
+         3f:40:5c:87:49:59:7e:88:6c:85:51:71:3b:a7:06:9f:97:d2:
+         ee:34:38:b7:c9:f3:98:64:b2:10:df:89:81:1f:84:6c:bf:29:
+         a3:f7:14:93:62:58:96:0c:41:14:fe:8e:d1:ae:b6:5e:12:a9:
+         0e:c4:d7:8b:87:ae:0b:6e:78:9f:76:b6:14:a6:eb:73:c8:15:
+         11:83:6c:6c:f6:1a:40:b3:2b:59:25:7f:74:88:95:bc:69:83:
+         cc:d5:a4:3b:79:eb:6a:4d:58:88:46:8a:94:02:62:10:a1:b9:
+         ca:94:d9:e9:3a:58:81:1e:5f:a9:45:8f:80:b5:39:ec:7c:07:
+         ad:29:43:fb:3a:c0:19:90:74:53:18:05:f1:d0:91:c9:43:52:
+         13:63:cc:39:0d:cc:3f:bf:56:c3:09:b0:1f:90:3b:9b:04:fd:
+         d7:fa:eb:3d:45:09:36:87:a9:bb:e2:15:85:14:db:bd:c4:e7:
+         e4:35:e8:cf:7f:b8:78:83:2c:1f:0e:3a:3f:c0:66:45:49:62:
+         b2:8b:a6:af:07:d5:79:df:39:d9:81:15:30:73:4d:3d:a7:93:
+         0c:f5:ed:fb:30:7f:29:c8:1a:9f:9d:b4:b0:3a:3e:21:52:8c:
+         9e:71:10:88:00:57:72:c2:33:1b:f5:a9:0f:da:ed:79:5c:fd:
+         cc:3a:d2:4e:12:f7:9c:c8:2c:4d:4d:90:71:ba:94:e0:eb:9c:
+         c1:e3:b1:9e:48:0b:9f:9d:cf:05:69:36:64:13:cc:4f:b8:2b:
+         5b:be:a3:95:3a:45:98:ab:39:6a:35:d0:9f:2b:c7:f7:01:97:
+         65:a7:02:e1:df:cc:1d:52:6e:46:ec:b2:2e:e9:a4:0d:89:8e:
+         ba:69:0b:3d:a9:5a:68:1a:fa:9d:9b:11:d1:92:e4:b0:ab:06:
+         71:1a:0c:dc:63:5c:89:69:61:97:43:a4:78:8e:00:b7:54:a9:
+         4c:4d:45:6c:d5:66:15:ea:3a:d8:2c:77:15:b7:a4:67:ff:61:
+         55:4c:e9:ec:dc:99:fa:71:33:ce:74:a5:62:62:8a:85:fb:25:
+         bc:c8:21:00:a8:0b:24:67
